refactor(vended-tools): give the sandbox-routed shell its own directory
Review feedback on the rename follow-up: the sandbox-routed shell tool lived
inside the bash directory, and the deprecated make_bash cluttered the Python
package __init__.

TypeScript now has vended-tools/shell/ holding makeShell, the Shell* errors,
and their tests, with a ./vended-tools/shell subpath export; vended-tools/bash/
keeps the persistent host bash tool and re-exports the shell names so imports
through the pre-rename path keep working until v2.0.0. The Shell* errors still
extend the Bash* errors, so the type dependency points from shell to bash and
pre-rename catch clauses keep matching.

Python moves make_bash and the rename rationale into shell/shell.py next to
the shell and bash instances; the package __init__ keeps only the __getattr__
that emits the DeprecationWarning for the bash instance.
