+++ strands-ts/src/sandbox/docker.ts
@@ -8,7 +8,7 @@ import { streamProcess } from './stream-process.js'
 import type { ExecutionResult, StreamChunk } from './types.js'
 import type { Tool } from '../tools/tool.js'
 import { makeFileEditor, DEFAULT_FILE_EDITOR_DESCRIPTION } from '../vended-tools/file-editor/index.js'
-import { makeShell, SANDBOX_SHELL_DESCRIPTION } from '../vended-tools/bash/index.js'
+import { makeShell, SANDBOX_SHELL_DESCRIPTION } from '../vended-tools/shell/index.js'
 
 /**
  * Options for constructing a {@link DockerSandbox}.
