+++ strands-ts/src/vended-tools/bash/types.ts
@@ -1,15 +1,9 @@
 /**
- * Type definitions for the bash and sandbox-routed shell tools.
+ * Type definitions for the persistent bash tool.
+ *
+ * The sandbox-routed shell tool's types live in `../shell/types.js`; its error
+ * classes extend the ones here so pre-rename catch clauses keep matching.
  */
-
-export const SANDBOX_SHELL_DESCRIPTION =
-  'Executes shell commands. Each call runs in a fresh shell; ' +
-  'state such as variables and the working directory does not persist across calls.'
-
-/**
- * @deprecated SANDBOX_BASH_DESCRIPTION is deprecated and will be removed in v2.0.0. Use SANDBOX_SHELL_DESCRIPTION instead. The tool routes commands through the sandbox, which runs sh or the remote login shell rather than bash specifically.
- */
-export const SANDBOX_BASH_DESCRIPTION = SANDBOX_SHELL_DESCRIPTION
 
 /**
  * Input parameters for execute operation.
@@ -85,31 +79,5 @@ export class BashSessionError extends Error {
   constructor(message: string) {
     super(message)
     this.name = 'BashSessionError'
-  }
-}
-
-/**
- * Error thrown when a sandbox-routed shell command exceeds its timeout.
- *
- * Extends {@link BashTimeoutError} so that callers who caught the previous error
- * type keep working; new code should catch this instead.
- */
-export class ShellTimeoutError extends BashTimeoutError {
-  constructor(message: string) {
-    super(message)
-    this.name = 'ShellTimeoutError'
-  }
-}
-
-/**
- * Error thrown when a sandbox-routed shell command fails.
- *
- * Extends {@link BashSessionError} so that callers who caught the previous error
- * type keep working; new code should catch this instead.
- */
-export class ShellExecutionError extends BashSessionError {
-  constructor(message: string) {
-    super(message)
-    this.name = 'ShellExecutionError'
   }
 }
